requirements.txt
pandas
numpy
Pillow
scikit-learn
joblib
